Create 01-how-can-i-secure-my-account-with-a-strong-password.md
Normally, we don't have periods at the end of lists.  I think that we should make an exception for this article.  It's a list, but they are also complete thoughts.  What say you?

This is copied directly from the document, so it should be ready to go except for the link.